requirements.txt
# Core ML/DL libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0

# SAM 2 and related
segment-anything-2>=0.1.0
groundingdino-py>=0.4.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
kornia>=0.6.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wandb>=0.15.0

# Jupyter and notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0

# Domain-specific
rasterio>=1.3.0  # Satellite imagery
fiona>=1.9.0     # Geospatial data
geopandas>=0.13.0  # Geospatial analysis

# Evaluation metrics
pycocotools>=2.0.6
timm>=0.9.0

# Optional: GPU acceleration
# cupy-cuda11x>=12.0.0  # Uncomment for CUDA 11.x
# cupy-cuda12x>=12.0.0  # Uncomment for CUDA 12.x 